Estable

+++ app/src/main/java/com/example/gaboq/instapoo/filters/DecompositionMin.java
@@ -21,7 +21,7 @@ public class DecompositionMin extends Imagen {
     @Override
     public void applyFilter() {
         for (int i = 0; i <this.length; i++) {
-            byte r,g,b;
+            int r,g,b;
             r = pixels[i].getR();
             g = pixels[i].getG();
             b = pixels[i].getB();
